Reorganising ui package
Adding activity, fragment, dialog and animation subpackages.
