+++ app/src/main/java/gc/david/dfm/dagger/FaqComponent.java
@@ -19,7 +19,7 @@ package gc.david.dfm.dagger;
 import javax.inject.Singleton;
 
 import dagger.Component;
-import gc.david.dfm.ui.HelpAndFeedbackActivity;
+import gc.david.dfm.ui.activity.HelpAndFeedbackActivity;
 
 /**
  * Created by david on 27.12.16.
