+++ app/src/main/java/gc/david/dfm/ui/animation/AnimatorUtil.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package gc.david.dfm.ui;
+package gc.david.dfm.ui.animation;
 
 import android.animation.Animator;
 import android.view.View;
@@ -23,11 +23,11 @@ import android.view.animation.DecelerateInterpolator;
 /**
  * Created by david on 09.01.17.
  */
-class AnimatorUtil {
+public class AnimatorUtil {
 
     private static final long ANIMATION_DURATION = 500L;
 
-    static void replaceViews(final View view1, final View view2) {
+    public static void replaceViews(final View view1, final View view2) {
         final int[] elevationChartLocation = new int[2];
         view1.getLocationOnScreen(elevationChartLocation);
