+++ app/src/main/java/gc/david/dfm/ui/fragment/SettingsFragment.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package gc.david.dfm.ui;
+package gc.david.dfm.ui.fragment;
 
 import android.os.Bundle;
 import android.support.v7.preference.Preference;
