+++ app/src/main/res/layout-land/activity_show_info.xml
@@ -22,7 +22,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     android:orientation="vertical"
-    tools:context=".ui.ShowInfoActivity">
+    tools:context=".ui.activity.ShowInfoActivity">
 
     <include layout="@layout/toolbar"/>
