+++ app/src/main/res/layout/activity_about.xml
@@ -21,7 +21,7 @@
               android:layout_height="match_parent"
               android:fitsSystemWindows="true"
               android:orientation="vertical"
-              tools:context=".ui.AboutActivity">
+              tools:context=".ui.activity.AboutActivity">
 
     <include
         android:id="@+id/tbMain"
