+++ app/src/main/res/layout/activity_help_and_feedback.xml
@@ -21,7 +21,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ui.HelpAndFeedbackActivity">
+    tools:context=".ui.activity.HelpAndFeedbackActivity">
 
     <include
         layout="@layout/toolbar"/>
